+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDaoIml.java
@@ -34,7 +34,7 @@ public class UserDaoIml implements UserDao {
     }
 
     @Override
-    public void editUser(int id, User user) {
+    public void editUser(Long id, User user) {
         entityManager.merge(user);
     }
