+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService {
 
     void removeUserById(Long id);
 
-    void editUser(int id, User user);
+    void editUser(Long id, User user);
 
     User showUser(Long id);
